Remove react-router references Correct window checking
