+++ packages/social-trader/src/hooks/location.ts
@@ -5,7 +5,7 @@ import { getLocation } from "utils/location";
 export const useLocation = () => {
   const [location, setLocation] = useState<Location | undefined>();
   useEffect(() => {
-    if (typeof window !== undefined) setLocation(getLocation());
+    if (typeof window !== "undefined") setLocation(getLocation());
   }, [window]);
   return { location };
 };
@@ -22,7 +22,7 @@ export const useParams = () => {
 export const useReferrer = () => {
   const [referrer, setReferrer] = useState<string | null>(null);
   useEffect(() => {
-    if (typeof window !== undefined) setReferrer(window.document.referrer);
+    if (typeof window !== "undefined") setReferrer(window.document.referrer);
   }, [window]);
   return referrer;
 };
