--- packages/social-trader/src/reducers/router-reducer.d.ts
@@ -1,11 +1,0 @@
-import { RouterActionType } from "connected-react-router";
-import { Location } from "history";
-
-interface LocationState extends Location {
-  prevPath?: string;
-}
-
-export interface RouterState {
-  location: LocationState;
-  action: RouterActionType;
-}
